Added arguments and code for link html wrappers
This added arguments and related code to allow the user to specify elements and classes to wrap the links. This allows for more control over the HTML displayed and the classes to be used for styling. I didn't change how it is displayed if you set the 'show_posts_links' varible so that  it only shows next and previous. That should come next.